Try fixing e2e tests
