+++ packages/e2e-tests/specs/experiments/multi-entity-editing.test.js
@@ -11,6 +11,7 @@ import {
 	visitAdminPage,
 	createNewPost,
 	publishPost,
+	trashAllPosts,
 } from '@wordpress/e2e-test-utils';
 import { addQueryArgs } from '@wordpress/url';
 
@@ -18,7 +19,6 @@ import { addQueryArgs } from '@wordpress/url';
  * Internal dependencies
  */
 import { useExperimentalFeatures } from '../../experimental-features';
-import { trashExistingPosts } from '../../config/setup-test-framework';
 
 const visitSiteEditor = async () => {
 	const query = addQueryArgs( '', {
@@ -164,8 +164,8 @@ describe( 'Multi-entity editor states', () => {
 	] );
 
 	beforeAll( async () => {
-		await trashExistingPosts( 'wp_template' );
-		await trashExistingPosts( 'wp_template_part' );
+		await trashAllPosts( 'wp_template' );
+		await trashAllPosts( 'wp_template_part' );
 	} );
 
 	it( 'should not display any dirty entities when loading the site editor', async () => {
@@ -194,8 +194,8 @@ describe( 'Multi-entity editor states', () => {
 
 	describe( 'Multi-entity edit', () => {
 		beforeAll( async () => {
-			await trashExistingPosts( 'wp_template' );
-			await trashExistingPosts( 'wp_template_part' );
+			await trashAllPosts( 'wp_template' );
+			await trashAllPosts( 'wp_template_part' );
 			await createNewPost( {
 				postType: 'wp_template',
 				title: kebabCase( templateName ),
